adding listen_in_background method

+++ src/pyssistant/sentence_analysis_module.py
@@ -10,20 +10,21 @@ def get_time_str():
     return time
 
 def play_message(sentence):
-    tts = gTTS(text=sentence, lang='en')
-    filename = "/tmp/temp" + get_time_str() + ".mp3"
-    tts.save(filename)
+    if sentence != '' or sentence != '\n':
+        tts = gTTS(text=sentence, lang='en')
+        filename = "/tmp/temp" + get_time_str() + ".mp3"
+        tts.save(filename)
 
-    playsound(filename)
+        playsound(filename)
 
-    os.remove(filename)
+        os.remove(filename)
 
 class SentenceAnalysis:
 
     def __init__(self, dialog_list):
         self.dialog_list = dialog_list
 
-    def get_main_command(self, result):
+    def get_command(self, result):
 
         message = ""
         conversation_state = True
@@ -34,6 +35,34 @@ class SentenceAnalysis:
         print "\n" + message + "\n"
         play_message(message)
         return conversation_state
+
+    def get_command_in_background(self, result):
+        result = result.lower()
+
+        if result != '':
+            message = self.__get_message_background(result)
+
+            print "\n" + message + "\n"
+            try:
+                play_message(message)
+            except Exception as e:
+                pass
+
+    def __get_message_background(self, result):
+        message = ""
+
+        for feature in self.dialog_list:
+            if feature.command in result:
+                if(feature.answer != None):
+                    message = feature.answer
+                if(feature.function != None):
+                    function = feature.function()
+                    if(type(function) == str):
+                        message = function
+                    elif(function != None):
+                        conversation_state = function
+
+        return message
 
     def __get_message(self, result):
         message = ""
